complicate lookups, simplify insertion
This also solves a bug when inserting a medium length prefix B between
(IP space) a short one C and a long one A, but after (temporal) the
others. In this case, queries more specific than B would be mapped to A,
as the less specific nodes between B and C would be mapped to A, but the
insertion of B cannot know whether the mapping to A is inherited or a
more specific prefix in it's own right (as C is). The set_for_subtrie
method removed by this commit was designed for the case when A is
inserted last and the inheritance needs to bridge the gap, but this is
unfeasible when the gap contains inherited, non-nil payloads.